+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ### [13.21.0](https://kaos.sh/ek/13.21.0)
 
 - **`[knf/united]`** Added methods `Is` and `Has`
+- **`[knf]`** Method `HasProp` renamed to `Has`
+- **`[knf]`** Code refactoring
 
 ### [13.20.5](https://kaos.sh/ek/13.20.5)